backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.5.2
python-dotenv>=1.0.0
opencv-python-headless>=4.8.1.78
numpy>=1.26.2
alembic>=1.12.1
bcrypt>=4.1.2
openpyxl>=3.1.2
google-auth>=2.25.2
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
google-generativeai>=0.3.2
firebase-admin>=6.4.0
